Refactoring, auth and checkRole middleware

+++ server/controllers/concertController.js
@@ -1,30 +1,26 @@
 // createConcert, deleteConcert, getConcerts (all, byType, byDate, byVenue), getOneConcertById
 const Concert = require('../schemas/concertSchema');
+const Ticket = require('../schemas/ticketSchema');
 
 const getConcerts = async(req, res) =>{
     try{
-        let concerts
+        let concerts = []
         let { type, date, venue } = req.query;
-        const query = {};
-        if (type) {
-            query.type = type;
-        }
-        if (date) {
-            query.date = date;
-        }
-        if (venue) {
-            query.venue = venue;
-        }
         if (!type && !date && !venue){
             concerts = await Concert.find();
-        }else {
-            console.log("Ok")
+        } else {
+            const query = {};
+            if (type) query.type = type;
+            if (date) query.date = date;
+            if (venue) query.venue = venue;
             concerts = await Concert.find(query);
         }
+        console.log(req.headers);
         return res.json(concerts);
     }
     catch (error){
         console.error(error);
+        return res.status(500);
     }
 };
 
@@ -36,6 +32,7 @@ const getOneConcertById = async(req, res) =>{
     }
     catch (error){
         console.error(error);
+        return res.status(500);
     }
     
 };
@@ -48,17 +45,23 @@ const createConcert = async(req, res) =>{
     }
     catch (error){
         console.error(error);
+        return res.status(500);
     }
 };
 
 const deleteConcert = async(req, res) =>{
     try{
         const { id } = req.params;
+        const concert = await Concert.findById(id)
+        if(!concert) {
+            return res.status(404).json(`There are no concerts with id ${id}`)
+        }
         await Concert.deleteOne({_id: id});
         return res.json("The concert has been removed");
     }
     catch (error){
         console.error(error);
+        return res.status(500);
     }
 };
 module.exports = {
